clenup respose types, clenup
